# contracts/openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: AI E-commerce Agent HTTP API
  version: 0.1.0
  description: |
    Authored, contract-first HTTP boundary for the local MVP-0 fixed workspace.
    This OpenAPI Description is the sole public contract authority. It describes
    resources and typed commands only; it does not imply that an HTTP handler,
    database, worker, model provider, or frontend has been implemented.
servers:
  - url: /
    description: Server-bound loopback same-origin workbench
...
paths:
  /api/v1/tasks:
    post:
      tags: [Task]
      operationId: createTask
      summary: Create a task without starting workflow execution
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            example:
              taskName: City commuter backpack launch
              productCategory: urban commuter backpack
              promotionGoal: clarify positioning for a spring launch
      responses:
        '201':
          description: Task created on the first submission
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskOverview'
        '200':
          description: Committed same-input idempotent replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskOverview'
        '400': { $ref: '#/components/responses/Problem400' }
        '409': { $ref: '#/components/responses/Problem409' }
        '413': { $ref: '#/components/responses/Problem413' }
        '415': { $ref: '#/components/responses/Problem415' }
        '422': { $ref: '#/components/responses/Problem422' }
        '500': { $ref: '#/components/responses/Problem500' }
        '503': { $ref: '#/components/responses/Problem503' }
    get:
      tags: [Task]
      operationId: listTasks
      summary: List a bounded recent-task window in the fixed workspace
      parameters:
        - $ref: '#/components/parameters/TaskLimit'
      responses:
        '200':
          description: Recent tasks (empty items is a normal response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummaryList'
        '400': { $ref: '#/components/responses/Problem400' }
        '500': { $ref: '#/components/responses/Problem500' }
        '503': { $ref: '#/components/responses/Problem503' }
  /api/v1/tasks/{taskId}:
    get:
      tags: [Task]
      operationId: getTaskOverview
      summary: Read the narrow task navigation overview
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskOverview'
        '404': { $ref: '#/components/responses/Problem404' }
        '500': { $ref: '#/components/responses/Problem500' }
        '503': { $ref: '#/components/responses/Problem503' }
  /api/v1/tasks/{taskId}/primary-input:
    get:
      tags: [Task]
      operationId: getTaskPrimaryInput
      summary: Read the current Task-scoped primary input
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Current primary input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPrimaryInput'
        '404': { $ref: '#/components/responses/Problem404' }
        '500': { $ref: '#/components/responses/Problem500' }
        '503': { $ref: '#/components/responses/Problem503' }
    put:
      tags: [Task]
      operationId: putTaskPrimaryInput
      summary: Replace the current Task-scoped primary input
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveTaskPrimaryInputRequest'
      responses:
        '200':
          description: Persisted current primary input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPrimaryInput'
        '404': { $ref: '#/components/responses/Problem404' }
        '409': { $ref: '#/components/responses/Problem409' }
        '413': { $ref: '#/components/responses/Problem413' }
        '422': { $ref: '#/components/responses/Problem422' }
        '500': { $ref: '#/components/responses/Problem500' }
        '503': { $ref: '#/components/responses/Problem503' }
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Project-defined retry identity scoped to the current fixed workspace and operation.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    TaskLimit:
      name: limit
      in: query
      required: false
      description: Server-bounded recent-task window size.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    Problem400:
      description: Malformed request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem404:
      description: Resource does not exist in the fixed workspace
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem409:
      description: Typed revision, idempotency, supersession, capability, or in-progress conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem413:
      description: Request or upload boundary is too large
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem415:
      description: Unsupported request media type
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem422:
      description: Semantically invalid or unprocessable input
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    Problem500:
      description: Safe internal error projection
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem503:
      description: Service unavailable at the HTTP boundary
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  headers:
    RetryAfter:
      description: Optional bounded retry hint; exact backoff is owned by RFC-007.
      schema:
        type: integer
        minimum: 0
    Location:
      description: Canonical resource or Run monitor URI.
      schema:
        type: string
        format: uri-reference
...
  schemas:
    ResourceReference:
      type: object
      required: [resourceKind, resourceId]
      properties:
        resourceKind:
          type: string
          minLength: 1
        resourceId:
          type: string
          minLength: 1
    MutableResourceReference:
      allOf:
        - $ref: '#/components/schemas/ResourceReference'
        - type: object
          required: [revision]
          properties:
            revision:
              $ref: '#/components/schemas/ResourceRevision'
    DomainVersionReference:
      type: object
      additionalProperties: false
      required: [resourceKind, resourceVersionId, versionNumber]
      properties:
        resourceKind:
          type: string
          minLength: 1
        resourceVersionId:
          type: string
          minLength: 1
        versionNumber:
          type: integer
          minimum: 1
    ResourceRevision:
      type: integer
      minimum: 0
...
    RunReference:
      type: object
      additionalProperties: false
      required: [runId]
      properties:
        runId:
          type: string
          minLength: 1
    StageReference:
      type: string
      enum:
        - product_intake_and_fact_extraction
        - customer_insight_analysis
        - product_positioning
        - human_review
        - marketing_brief_generation
        - xiaohongshu_brief_mapping
    StageStatus:
      type: string
      enum: [not_started, ready, running, waiting_input, waiting_review, valid, invalid, failed, skipped]
    StageSummary:
      type: object
      additionalProperties: false
      required: [stage, status, currentVersion, lastValidVersion, lastRun, waitingReason, updatedAt]
      properties:
        stage: { $ref: '#/components/schemas/StageReference' }
        status: { $ref: '#/components/schemas/StageStatus' }
        currentVersion:
          oneOf:
            - $ref: '#/components/schemas/DomainVersionReference'
            - type: 'null'
        lastValidVersion:
          oneOf:
            - $ref: '#/components/schemas/DomainVersionReference'
            - type: 'null'
        lastRun:
          oneOf:
            - $ref: '#/components/schemas/RunReference'
            - type: 'null'
        waitingReason:
          oneOf:
            - type: string
            - type: 'null'
        updatedAt:
          type: string
          format: date-time
    TaskStatus:
      type: string
      enum: [draft, running, waiting_for_input, waiting_for_review, paused, completed, failed, cancelled]
...
    Capability:
      type: string
      enum:
        - start
        - cancel
        - resume
        - rerun
        - retry_current_stage
        - restart_from_safe_boundary
        - resolve_needs_input
        - preview_remove_source
        - remove_source
        - preview_replace_source
        - replace_source
        - save_review_draft
        - submit_review
        - request_more_information
        - reject_all_and_request_regeneration
        - withdraw_approved_strategy
        - compare_brief
        - revise_marketing_brief
        - revise_xiaohongshu_brief
        - preview_export
        - confirm_export
    PrimaryAction:
      oneOf:
        - $ref: '#/components/schemas/NoPrimaryAction'
        - $ref: '#/components/schemas/NavigatePrimaryAction'
        - $ref: '#/components/schemas/CommandPrimaryAction'
      discriminator:
        propertyName: type
    NoPrimaryAction:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          const: none
    NavigatePrimaryAction:
      type: object
      additionalProperties: false
      required: [type, target]
      properties:
        type:
          const: navigate
        target:
          type: string
          enum: [intake, needs_input, review, results, recovery]
    CommandPrimaryAction:
      type: object
      additionalProperties: false
      required: [type, command]
      properties:
        type:
          const: command
        command: { $ref: '#/components/schemas/Capability' }
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required: [taskName, productCategory, promotionGoal]
      properties:
        taskName:
          type: string
          minLength: 1
        productCategory:
          type: string
          minLength: 1
        promotionGoal:
          type: string
          minLength: 1
    PrimaryInputKind:
      type: string
      enum: [pasted_text, text_file, markdown_file]
    SaveTaskPrimaryInputRequest:
      type: object
      additionalProperties: false
      required: [inputKind, fileName, content]
      properties:
        inputKind: { $ref: '#/components/schemas/PrimaryInputKind' }
        fileName:
          oneOf:
            - type: string
              minLength: 1
            - type: 'null'
          description: Display basename for a text/markdown file; null for pasted text.
        content:
          type: string
          minLength: 1
          description: Nonblank UTF-8 text, limited to 1 MiB after UTF-8 encoding.
    TaskPrimaryInput:
      type: object
      additionalProperties: false
      required: [taskId, inputRevision, inputKind, fileName, content, byteCount, updatedAt]
      properties:
        taskId:
          type: string
          minLength: 1
        inputRevision: { $ref: '#/components/schemas/ResourceRevision' }
        inputKind: { $ref: '#/components/schemas/PrimaryInputKind' }
        fileName:
          oneOf:
            - type: string
              minLength: 1
            - type: 'null'
        content:
          type: string
          minLength: 1
        byteCount:
          type: integer
          minimum: 1
          maximum: 1048576
        updatedAt:
          type: string
          format: date-time
    TaskSummary:
      type: object
      required: [taskId, taskName, productCategory, taskStatus, currentStage, waitingReason, updatedAt, revision, primaryAction, capabilities]
      properties:
        taskId:
          type: string
          minLength: 1
        taskName:
          type: string
          minLength: 1
        productCategory:
          type: string
          minLength: 1
        taskStatus: { $ref: '#/components/schemas/TaskStatus' }
        currentStage:
          oneOf:
            - $ref: '#/components/schemas/StageReference'
            - type: 'null'
        waitingReason:
          oneOf:
            - type: string
            - type: 'null'
        updatedAt:
          type: string
          format: date-time
        revision: { $ref: '#/components/schemas/ResourceRevision' }
        primaryAction: { $ref: '#/components/schemas/PrimaryAction' }
        capabilities:
          type: array
          items: { $ref: '#/components/schemas/Capability' }
          uniqueItems: true
    TaskSummaryList:
      type: object
      additionalProperties: false
      required: [items, limit]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TaskSummary' }
        limit:
          type: integer
          minimum: 1
          maximum: 50
    TaskOverview:
      allOf:
        - $ref: '#/components/schemas/TaskSummary'
        - type: object
          additionalProperties: false
          required: [stages, activeRun, latestRun, needsInputRequest, reviewPackage, approvedStrategy, marketingBrief, xiaohongshuBrief]
          properties:
            stages:
              type: array
              items: { $ref: '#/components/schemas/StageSummary' }
            activeRun:
              oneOf:
                - $ref: '#/components/schemas/RunReference'
                - type: 'null'
            latestRun:
              oneOf:
                - $ref: '#/components/schemas/RunReference'
                - type: 'null'
            needsInputRequest:
              oneOf:
                - $ref: '#/components/schemas/MutableResourceReference'
                - type: 'null'
            reviewPackage:
              oneOf:
                - $ref: '#/components/schemas/ReviewPackageReference'
                - type: 'null'
            approvedStrategy:
              oneOf:
                - $ref: '#/components/schemas/DomainVersionReference'
                - type: 'null'
            marketingBrief:
              oneOf:
                - $ref: '#/components/schemas/DomainVersionReference'
                - type: 'null'
            xiaohongshuBrief:
              oneOf:
                - $ref: '#/components/schemas/DomainVersionReference'
                - type: 'null'
...
    ReviewPackageReference:
      type: object
      additionalProperties: false
      required: [reviewPackageId, packageVersion]
      properties:
        reviewPackageId:
          type: string
          minLength: 1
        packageVersion:
          type: integer
          minimum: 1
...
    ProblemAction:
      type: string
      enum: [correct_input, refresh, refresh_and_compare, open_current, retry_later, contact_operator, none]
    ProblemType:
      type: string
      enum:
        - urn:ai-ecommerce-agent:problem:malformed-request
        - urn:ai-ecommerce-agent:problem:not-found
        - urn:ai-ecommerce-agent:problem:payload-too-large
        - urn:ai-ecommerce-agent:problem:unsupported-media-type
        - urn:ai-ecommerce-agent:problem:validation-failed
        - urn:ai-ecommerce-agent:problem:revision-conflict
        - urn:ai-ecommerce-agent:problem:idempotency-conflict
        - urn:ai-ecommerce-agent:problem:superseded-resource
        - urn:ai-ecommerce-agent:problem:capability-conflict
        - urn:ai-ecommerce-agent:problem:operation-in-progress
        - urn:ai-ecommerce-agent:problem:rate-limited
        - urn:ai-ecommerce-agent:problem:internal-error
        - urn:ai-ecommerce-agent:problem:service-unavailable
    FieldIssue:
      type: object
      additionalProperties: false
      required: [fieldPath, reasonCode]
      properties:
        fieldPath:
          type: string
          minLength: 1
        reasonCode:
          type: string
          minLength: 1
    ProblemConflictContext:
      type: object
      additionalProperties: false
      properties:
        currentResource:
          oneOf:
            - $ref: '#/components/schemas/ResourceReference'
            - type: 'null'
        currentRevision:
          oneOf:
            - $ref: '#/components/schemas/ResourceRevision'
            - type: 'null'
        currentVersion:
          oneOf:
            - $ref: '#/components/schemas/DomainVersionReference'
            - type: 'null'
        fieldPath:
          oneOf:
            - type: string
            - type: 'null'
        basisSummary:
          oneOf:
            - type: string
            - type: 'null'
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, action]
      example:
        type: urn:ai-ecommerce-agent:problem:revision-conflict
        title: Revision conflict
        status: 409
        detail: The resource changed; refresh and compare before retrying.
        instance: /api/v1/review-packages/review-package-01/draft
        action: refresh_and_compare
      properties:
        type: { $ref: '#/components/schemas/ProblemType' }
        title:
          type: string
          minLength: 1
        status:
          type: integer
          enum: [400, 404, 409, 413, 415, 422, 429, 500, 503]
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        action: { $ref: '#/components/schemas/ProblemAction' }
        fieldIssues:
          type: array
          items: { $ref: '#/components/schemas/FieldIssue' }
        conflict:
          $ref: '#/components/schemas/ProblemConflictContext'
        correlationReference:
          type: string
          minLength: 1
        retryAfterSeconds:
          type: integer
          minimum: 0